+++ libspider/libspider/spider.cpp
@@ -94,7 +94,7 @@ static std::vector<spider::pisdf::Refinement *> specialActorRefinements() {
 
 /* === Function(s) definition === */
 
-void spider::api::createGenericStack(StackID stack, const std::string &name, std::int32_t alignment) {
+void spider::api::createGenericStack(StackID stack, const std::string &name, std::uint64_t alignment) {
     createAllocator(spider::type<AllocatorType::GENERIC>{ }, stack, name, alignment);
 }
 
@@ -102,7 +102,7 @@ void spider::api::createFreeListStack(StackID stack,
                                       const std::string &name,
                                       std::uint64_t staticBufferSize,
                                       FreeListPolicy policy,
-                                      std::int32_t alignment) {
+                                      std::uint64_t alignment) {
     createAllocator(spider::type<AllocatorType::FREELIST>{ }, stack, name, staticBufferSize, policy, alignment);
 }
 
@@ -110,7 +110,7 @@ void spider::api::createFreeListStaticStack(StackID stack,
                                             const std::string &name,
                                             std::uint64_t totalSize,
                                             FreeListPolicy policy,
-                                            std::int32_t alignment) {
+                                            std::uint64_t alignment) {
     createAllocator(spider::type<AllocatorType::FREELIST_STATIC>{ }, stack, name, totalSize, policy, alignment);
 }
 
@@ -119,14 +119,14 @@ void spider::api::createFreeListStaticStack(StackID stack,
                                             std::uint64_t totalSize,
                                             void *base,
                                             FreeListPolicy policy,
-                                            std::int32_t alignment) {
+                                            std::uint64_t alignment) {
     createAllocator(spider::type<AllocatorType::FREELIST_STATIC>{ }, stack, name, totalSize, base, policy, alignment);
 }
 
 void spider::api::createLinearStaticStack(StackID stack,
                                           const std::string &name,
                                           std::uint64_t totalSize,
-                                          std::int32_t alignment) {
+                                          std::uint64_t alignment) {
     createAllocator(spider::type<AllocatorType::LINEAR_STATIC>{ }, stack, name, totalSize, alignment);
 }
 
@@ -134,7 +134,7 @@ void spider::api::createLinearStaticStack(StackID stack,
                                           const std::string &name,
                                           std::uint64_t totalSize,
                                           void *base,
-                                          std::int32_t alignment) {
+                                          std::uint64_t alignment) {
     createAllocator(spider::type<AllocatorType::LINEAR_STATIC>{ }, stack, name, totalSize, base, alignment);
 }
